meta-ampere: mtmitchell: change platform name
There is a major change in Mt.Mitchell hardware from EVT to DVT version
and we need to support the EVT version for a while. As the platform name
for Mt.Mitchell EVT was mtmitchell, we used mtmitchell-dcscm platform
name for DVT and later hardware version.
It is the time to drop EVT support now. But it is high impact with
multiple stake holders so will keep mtmitchell-dcscm from now on.
This commit changes platform name to mtmitchell-dcscm to indicate this.

Signed-off-by: Thang Q. Nguyen <thang@os.amperecomputing.com>
Change-Id: Iea7a73f9df83ae01526b94c32173bfca0707935d