meta-ampere: PSM: support BMC state UpdateInProgress value
As defined in the PDIs [1], the BMC's state is UpdateInProgress while
BMC is in Update Mode, but the phosphor-bmc-state-manager doesn't handle
this state.

The phosphor-software-manager is handling update actions, when users
request to update bios/bmc ..., the phosphor-software-manager creates
ActivationBlocksTransition interface [2] to block transition actions.

This commit adds monitoring to ActivationBlocksTransition interface. If
this interface is added, the BMC's state should be updated to
UpdateInProgress value. If this interface is removed, the BMC's state
should be updated to Ready.

[1] https://github.com/openbmc/phosphor-dbus-interfaces/blob/855b46f32ad385b9e26cf3a2c1e3a547e0eb8632/yaml/xyz/openbmc_project/State/BMC.interface.yaml#L57
[2] https://github.com/openbmc/phosphor-dbus-interfaces/blob/a1347418307d31a94bd21f22897aa508df938dcf/yaml/xyz/openbmc_project/Software/ActivationBlocksTransition.interface.yaml

Tested:
    1. Check the BMC's state.
       $busctl get-property xyz.openbmc_project.State.BMC \
               /xyz/openbmc_project/state/bmc0 \
               xyz.openbmc_project.State.BMC CurrentBMCState
    2. The BMC's state is "Ready".
    3. Request to flash new BIOS via Redfish/WebUI.
    4. Check the BMC's state as step 1.
    5. The BMC's state is "UpdateInProgress".
    6. Wait until updating finished.
    7. Check the BMC's state as step 1.
    8. The BMC's state is "Ready".

Signed-off-by: Thang Tran <thuutran@amperecomputing.com>