meta-ampere: flash: Update method to turn on the host
While BMC is in firmware updating phase, the BMC's state is
UpdateInProgress, BMC rejects all turn on host requests during this time
if users enabled "only-allow-boot-when-bmc-ready" option build.

This commit updates the method to turn on the host after flashing done.
When flashing done, BMC does not directly request to turn on the host,
it request to start turn-on-the-host-after-flash@.service instead. This
service shall check BMC's state then request to turn on the host. B/c
this service's "Type" is simple, it does not block firmware updating
action.

Tested:
    1. Turn on the Host then request to flashing new firmware.
    2. Wait until flashing done, check the Host's state.
       $busctl get-property xyz.openbmc_project.State.Host0
               /xyz/openbmc_project/state/host0 \
               xyz.openbmc_project.State.Host CurrentHostState
    3. The Host's state is "Running"

Signed-off-by: Thang Tran <thuutran@amperecomputing.com>