meta-ampere: led: Update priority property
Follow the LED priority configuration instruction [1], Each LED can
have "Priority" as "Blink", "Off" or "On". If this property is defined,
it should be defined on each instance of the LED in the config.

This commit updates the "Priority" property to all LEDs.

[1] https://github.com/openbmc/phosphor-led-manager/blob/master/README.md#configuration-led-priority

Tested: Check the xyz.openbmc_project.LED.GroupManager.service starts
successfully.

Signed-off-by: Chanh Nguyen <chanh@os.amperecomputing.com>