meta-ampere: Avoid restore of fan control in manual mode
As the ampere_fanctrl.sh script only stop the fan control services, in
case of a power cycle, all of them will be started due to their
dependencies with obmc-chassis-poweron@0.target, This commit enlists the
MaskUnitFiles and UnmaskUnitFiles methods on DBus instead to mask the
services so they can't be started when user disables fan control mode.

Moreover, when the fan control mode is enabled during chassis-off
condition, phopsphor-pid-control.service will only be unmasked and not
started to address the dependency of the service with chassis state.

Tested:

1. Host is ON
   $ ampere_fanctrl.sh setstatus 1 # to disable fan control mode
=> The fan control services are masked and stopped. Unless BMC is
rebooted or the fan control mode is enabled, the services can not
be started by any mean
2. $ ipmitool chassis power off
3. $ ampere_fanctrl.sh setstatus 0 # to enable fan control mode
=> The fan control services are un-masked and started, while
phosphor-pid-control.service is only unmasked.
4. $ ampere_fanctrl.sh setstatus 1 # to disable fan control mode
5. $ ipmitool chassis power on
=> The fan control services are not restarted unless the control mode is
enabled again
6. $ ampere_fanctrl.sh setstatus 0
=> The fan control services are un-masked and started, every works
normally

Signed-off-by: Chau Ly <chaul@amperecomputing.com>