meta-ampere: ampere-usbnet: Ignore link carrier state
The virtual LAN over USB (usb0) connection between BMC and HOST is
locally utilized by the Redfish Host Interface. It is configured with a
static IPv4 address '192.168.0.10' by the network configuration file [1]
'00-bmc-usb0.network'.

Returning to the Ampere Altra Mt. Jade platform, an AC power cycle
caused the usb0 IP address to be lost [2]. The `ipmi lan print 3`
command gives a zero IP address on which HOST has not yet initiated the
PCIe USB device, which was caused by the usb0 link not being ready. The
solution was determined and well-tested. However, the same issue with
the upstream driver in edk2 is still present today. Therefore, a clean
solution ought to be used to fix it.

This patch sets ConfigureWithoutCarrier to 'yes' to allows usb0
interface to retain the static IP address even if the physical carrier
link is down. This helps HOST UEFI firmware to retrieve the IP address
information using the `IPMI lan print 3` command to populate the SMBIOS
Type 42 for the subsequent redfish discovery even when the usb0 link is
not yet established.

Tested with the following steps:
  1. Check the usb0 IP address by the command `ip a`. The IP
     192.168.0.10 should be shown in the command output.
  2. Check the `ipmi lan print 3`. Since the IP address is available,
     the IP address shown in the command output should be idential with
     the one in step 1.
  3. `networkctl down usb0` - Set the usb0 link down.
  4. Run again step 1 & 2. The usb0 IP address should be retained.
     Without this patch, it returns all zeros.

[1] https://www.freedesktop.org/software/systemd/man/latest/systemd.network.html
[2] https://den-bugzilla.amperecomputing.com/show_bug.cgi?id=92830

Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>