meta-ampere: webui: fix for KVM terminal not show up
During the upgrade to Vue 3 [1], the initialization of the selector,
including the definitions of height, display type, background, overflow,
etc., was ignored, resulting in the KVM screen not displaying the
terminal screen. This commit helps to import these properties for the
elements mentioned above from the SCSS file.

Tested:
1. Go to the OBMC web page.
2. Go to the KVM page and check that it works normally
3. By minifying and zooming in on the web browser, the KVM screen
adjusts automatically.
4. Switch between different pages.
5. Return to the KVM page and check that it works normally.

1. https://github.com/openbmc/webui-vue/blob/master/src/App.vue#L38

Signed-off-by: HuyLe <hule@amperecomputing.com>