meta-ampere: host-ipmid: report correct IP address via configured IP source
In case both DHCP and static IP addresses are assigned, ipmi tool
just returns one of them. IPMI code always returns the first
object dbus on the interface by the fixed index, which is
always 0. That may cause a mismatch between the IP address and
configured IP address source which are either DHCP or static..

This patch will check if the DHCP or static mode is enabled,
then return the IP address accordingly.

This patch is being in progress at gerrit [1] for reviewing.

Tested:
1. Check DHCP IP address
   ipmitool lan print 1 | grep "IP Address"
     IP Address Source       : DHCP Address
     IP Address              : 10.39.122.87

2. Change to static source
   ipmitool lan set 1 ipsrc static
   ipmitool lan print 1 | grep "IP Address"
     IP Address Source       : Static Address
     IP Address              : 0.0.0.0

3. Set a static IP address.
   ipmitool lan set 1 ipaddr 10.39.122.50
   ipmitool lan print 1 | grep "IP Address"
     IP Address Source       : Static Address
     IP Address              : 10.39.122.50

4. Back to DHCP source
   Expect report DHCP IP address accordingly.

   ipmitool lan set 1 ipsrc dhcp
   ipmitool lan print 1 | grep "IP Address"
     IP Address Source       : DHCP Address
     IP Address              : 10.39.122.87

[1] https://gerrit.openbmc.org/c/openbmc/phosphor-host-ipmid/+/75745

Signed-off-by: Chanh Nguyen <chanh@os.amperecomputing.com>