meta-ampere: phosphor-fan: remove ExecStopPost action for Mt.Mitchell and Mt.Jefferson
As AC04 and AC04_1 don't have requirements to set default fan speed
when fan control services are stopped, this commit removes the behavior
of setting fans to 100% in ExecStopPost of phosphor-fan services. The
behavior in AC03 is still kept as is to ensure fans run 100% PWM
when BMC is rebooted. For AC03, this feature is implemented in
phosphor-fan-control@.service only..

Tested: In Mt.Jefferson and Mt.Mitchelll

1. When BMC is ON, host is ON, disable fan control
$ ampere_fanctrl.sh setstatus 1
=> Fans keep their current speed.

Signed-off-by: Chaul Ly <chaul@amperecomputing.com>